fix(domain): enforce grocery session intake lifecycle, delete protections, signed deltas, and decimal currency

+++ app/models/session.py
@@ -1,4 +1,5 @@
 import enum
+from decimal import Decimal
 from datetime import datetime
 from typing import Optional, List, TYPE_CHECKING
 from sqlalchemy import String, Enum as SQLEnum, DateTime, Numeric, Text, func
@@ -21,7 +22,7 @@ class GrocerySession(Base, TimestampMixin):
     id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
     store_name: Mapped[str] = mapped_column(String(255), nullable=False)
     purchase_date: Mapped[datetime] = mapped_column(DateTime(timezone=True), default=func.now(), nullable=False)
-    total_amount: Mapped[Optional[float]] = mapped_column(Numeric(10, 2), nullable=True)
+    total_amount: Mapped[Optional[Decimal]] = mapped_column(Numeric(10, 2), nullable=True)
     status: Mapped[GrocerySessionStatus] = mapped_column(
         SQLEnum(GrocerySessionStatus, name="grocery_session_status_enum"),
         default=GrocerySessionStatus.DRAFT,
